+++ src/App.jsx
@@ -1,5 +1,5 @@
 import { ScrollProgress } from "./components/scroll-progres/component";
-import { ThemeContextProvider } from "./components/theme-context-provider/component";
+import { ThemeContextProvider } from "./components/theme-context/provider";
 import { Header } from "./components/header/component";
 import { Footer } from "./components/footer/component";
 import { Restaurant } from "./components/restaurant/component";
